==> Data_preprocessing/0 Data preprocessing_vital_lab.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm
import re

feature_ID_list = pd.read_csv('./Feature ID list for shock(MIMIC-4).csv')
feature_name = feature_ID_list['Feature name'].tolist()
feature_ID_list.set_index('Feature name', inplace=True)

chartevents = pd.read_csv('./chartevents.csv')

#weight
feature_id_ls = feature_ID_list['chartevents'][feature_name[0]]
id_list = [int(x) for x in feature_id_ls.split(',')]
chart_total = chartevents[chartevents['itemid'].isin(id_list)]
chart_total['Feature name'] = feature_name[0]
chart_total = chart_total[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition1 = chart_total['valuenum']>=20 
chart_total = chart_total[condition1]
print(feature_name[0])

#height
feature_id_ls = feature_ID_list['chartevents'][feature_name[1]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Height'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
height_inch_ls = [226707]
height_inch_index = feature[feature['itemid'].isin(height_inch_ls)].index
height_inch = feature.loc[height_inch_index]
height_inch = height_inch.dropna(subset=['valuenum'])
height_inch_index = height_inch.index
feature['valuenum'][height_inch_index] = feature['valuenum'][height_inch_index].astype(float)
feature['valuenum'][height_inch_index] = feature['valuenum'][height_inch_index]*2.54
condition3 = feature['valuenum']<=250
condition1 = feature['valuenum']>=100
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#HR
feature_id_ls = feature_ID_list['chartevents'][feature_name[2]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Heart rate'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<300
condition1 = feature['valuenum']>10   
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#NISBP
feature_id_ls = feature_ID_list['chartevents'][feature_name[3]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'NIBlood pressure systolic'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=300
condition1 = feature['valuenum']>30
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#NIDBP
feature_id_ls = feature_ID_list['chartevents'][feature_name[4]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'NIBlood pressure diastolic'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=175
condition1 = feature['valuenum']>=10  
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#ISBP
feature_id_ls = feature_ID_list['chartevents'][feature_name[5]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'IBlood pressure systolic'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=300
condition1 = feature['valuenum']>30
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#IDBP
feature_id_ls = feature_ID_list['chartevents'][feature_name[6]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'IBlood pressure diastolic'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=175
condition1 = feature['valuenum']>=10  
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#RR
feature_id_ls = feature_ID_list['chartevents'][feature_name[7]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Respiratory rate'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=60
condition1 = feature['valuenum']>2 
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#SpO2
feature_id_ls = feature_ID_list['chartevents'][feature_name[8]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'SpO2'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=100
condition1 = feature['valuenum']>=10   
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#Temperature
feature_id_ls = feature_ID_list['chartevents'][feature_name[9]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Temperature'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
tem_F_ls = [223761]
tem_F_index = feature[feature['itemid'].isin(tem_F_ls)].index
tem_F = feature.loc[tem_F_index]
tem_F = tem_F.dropna(subset=['valuenum'])
tem_F_index = tem_F.index
feature['valuenum'][tem_F_index] = feature['valuenum'][tem_F_index].astype(float)
feature['valuenum'][tem_F_index] = (feature['valuenum'][tem_F_index]-32)/1.8
condition3 = feature['valuenum']<43
condition1 = feature['valuenum']>=29
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#Hemoglobin
feature_id_ls = feature_ID_list['chartevents'][feature_name[10]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Hemoglobin'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=20
condition1 = feature['valuenum']>=2
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#Creatinine
feature_id_ls = feature_ID_list['chartevents'][feature_name[11]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Creatinine'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=30
condition1 = feature['valuenum']>0.09 
feature = feature[condition1 & condition3]
chart_total = pd.concat([chart_total, feature], axis=0)

#Troponin
feature_id_ls = feature_ID_list['chartevents'][feature_name[12]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Troponin'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition1 = feature['valuenum']>=0.01  
feature = feature[condition1]
chart_total = pd.concat([chart_total, feature], axis=0)

#Lactate
feature_id_ls = feature_ID_list['chartevents'][feature_name[13]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = chartevents[chartevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Lactate'
feature = feature[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition1 = feature['valuenum']>=0.09  
feature = feature[condition1]
chart_total = pd.concat([chart_total, feature], axis=0)


chart_total = chart_total.sort_values(by=['subject_id', 'hadm_id', 'stay_id', 'charttime'], ascending=[True,True,True,True])
chart_total.reset_index(drop=True, inplace=True)

chart_total.to_csv('./pivoted chart table.csv', index=False)

chart_total = 0
chartevents = 0
labevents = pd.read_csv('./labevents.csv')

#SpO2
feature_id_ls = feature_ID_list['labevents'][feature_name[8]]
id_list = [int(x) for x in feature_id_ls.split(',')]
lab_total = labevents[labevents['itemid'].isin(id_list)]
lab_total['Feature name'] = 'SpO2'
lab_total = lab_total[['subject_id', 'hadm_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = lab_total['valuenum']<=100
condition1 = lab_total['valuenum']>=10  
lab_total = lab_total[condition1 & condition3]

#Temperature
feature_id_ls = feature_ID_list['labevents'][feature_name[9]]
id_list = [int(x) for x in feature_id_ls.split(',')]
lab_total = labevents[labevents['itemid'].isin(id_list)]
lab_total['Feature name'] = 'Temperature'
lab_total = lab_total[['subject_id', 'hadm_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = lab_total['valuenum']<43
condition1 = lab_total['valuenum']>=29  
lab_total = lab_total[condition1 & condition3]

#Hemoglobin
feature_id_ls = feature_ID_list['labevents'][feature_name[10]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = labevents[labevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Hemoglobin'
feature = feature[['subject_id', 'hadm_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=20
condition1 = feature['valuenum']>=2   
feature = feature[condition1 & condition3]
lab_total = pd.concat([lab_total, feature], axis=0)

#Creatinine
feature_id_ls = feature_ID_list['labevents'][feature_name[11]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = labevents[labevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Creatinine'
feature = feature[['subject_id', 'hadm_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition3 = feature['valuenum']<=30
condition1 = feature['valuenum']>0.09  
feature = feature[condition1 & condition3]
lab_total = pd.concat([lab_total, feature], axis=0)

#Tropinin
feature_id_ls = feature_ID_list['labevents'][feature_name[12]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = labevents[labevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Tropinin'
feature = feature[['subject_id', 'hadm_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition1 = feature['valuenum']>=0.01
feature = feature[condition1]
lab_total = pd.concat([lab_total, feature], axis=0)

#Lactate
feature_id_ls = feature_ID_list['labevents'][feature_name[13]]
id_list = [int(x) for x in feature_id_ls.split(',')]
feature = labevents[labevents['itemid'].isin(id_list)]
feature['Feature name'] = 'Lactate'
feature = feature[['subject_id', 'hadm_id', 'Feature name', 'itemid', 'charttime','valuenum']]
condition1 = feature['valuenum']>0.09
feature = feature[condition1]
lab_total = pd.concat([lab_total, feature], axis=0)

lab_total['stay_id'] = None
lab_total = lab_total[['subject_id', 'hadm_id', 'stay_id', 'Feature name', 'itemid', 'charttime','valuenum']]

lab_total = lab_total.sort_values(by=['subject_id', 'hadm_id', 'charttime'], ascending=[True,True,True])
lab_total.reset_index(drop=True, inplace=True)

lab_total.to_csv('./pivoted lab table.csv', index=False)

import pandas as pd
from tqdm.notebook import tqdm
import re

chart_total = pd.read_csv('./pivoted chart table.csv')
lab_total = pd.read_csv('./pivoted lab table.csv')

feature_ID_list = pd.read_csv('./Feature ID list for shock(MIMIC-4).csv')
feature_name = feature_ID_list['Feature name'].tolist()
feature_ID_list.set_index('Feature name', inplace=True)

main = chart_total
head = ['subject_id', 'hadm_id', 'stay_id', 'charttime']
new_df = pd.DataFrame(columns=head, index=range(len(main)))

for name in feature_name[:14]:
    sdata = main[main['Feature name']==name]
    feature_name_index = sdata.index
    new_df[name] = None
    new_df.loc[feature_name_index] = sdata[['subject_id', 'hadm_id', 'stay_id', 'charttime']]
    new_df[name].loc[feature_name_index] = sdata['valuenum']
    print(name)

print('------------------------------------------------------------------')

new_df = new_df.sort_values(by=['subject_id', 'hadm_id', 'stay_id', 'charttime'], ascending=[True,True,True,True])

new_df.reset_index(drop=True, inplace=True)

new_df.to_csv('./pivoted chart table.csv', index=False)

print(new_df.shape)
new_df.isnull().sum()

main = lab_total
head = ['subject_id', 'hadm_id', 'stay_id', 'charttime']
new_df = pd.DataFrame(columns=head, index=range(len(main)))

for name in feature_name[8:14]:
    sdata = main[main['Feature name']==name]
    feature_name_index = sdata.index
    new_df[name] = None
    new_df.loc[feature_name_index] = sdata[['subject_id', 'hadm_id', 'stay_id', 'charttime']]
    new_df[name].loc[feature_name_index] = sdata['valuenum']
    print(name)

print('------------------------------------------------------------------')

new_df = new_df.sort_values(by=['subject_id', 'hadm_id', 'charttime'], ascending=[True,True,True])

new_df.reset_index(drop=True, inplace=True)

new_df.to_csv('./pivoted lab table.csv', index=False)

%reset -f

import pandas as pd
from tqdm.notebook import tqdm
import re

feature_ID_list = pd.read_csv('./Feature ID list for shock(MIMIC-4).csv')
feature_name = feature_ID_list['Feature name'].tolist()
feature_ID_list.set_index('Feature name', inplace=True)

data = pd.read_csv('./pivoted lab table.csv')
data = data.dropna(subset=['hadm_id'])
data.drop(columns=['stay_id'], inplace=True)
patients = pd.read_csv('./icustays.csv')
patients.drop(columns=['stay_id'], inplace=True)
#patients.drop(columns=['Unnamed: 0'], inplace=True)
#patients.drop(columns=['ICUSTAY_ID'], inplace=True)

data = pd.merge(data, patients, on=['subject_id','hadm_id'], how='left')
#data = data[data['los']>1]
data.reset_index(drop=True, inplace=True)

data['charttime'] = pd.to_datetime(data['charttime'])
data['outtime'] = pd.to_datetime(data['outtime'])
data['intime'] = pd.to_datetime(data['intime'])

condition1 = (data['intime'] > data['charttime'])
condition2 = (data['intime'] - pd.to_timedelta(24, unit='h') < data['charttime'])

before_icu_data = data[condition1 & condition2]
before_icu_data = before_icu_data.sort_values(by=['subject_id', 'hadm_id', 'charttime'], ascending=[True,True,True])
before_icu_data.reset_index(drop=True, inplace=True)

before_icu_data.fillna(method='ffill', inplace=True)

def latest_data(x):
    x.reset_index(drop=True, inplace=True)
    df = x.loc[len(x)-1]

    return df

before_icu_data_latest = before_icu_data.groupby('hadm_id').apply(latest_data)
before_icu_data_latest.reset_index(drop=True, inplace=True)

bidl_ls = before_icu_data_latest['hadm_id'].tolist()

data = pd.read_csv('./pivoted chart table.csv')
data = data.dropna(subset=['hadm_id'])
patients = pd.read_csv('./icustays.csv')
#patients.drop(columns=['Unnamed: 0'], inplace=True)
#patients.drop(columns=['ICUSTAY_ID'], inplace=True)

data = pd.merge(data, patients, on=['subject_id','hadm_id','stay_id'], how='left')
#data = data[data['los']>1]
data.reset_index(drop=True, inplace=True)

print(data.shape)
data.isnull().sum()

condition3 = (data['intime'] <= data['charttime'])
condition4 = (data['outtime'] >= data['charttime'])

after_icu_data = data[condition3 & condition4]
after_icu_data = after_icu_data.sort_values(by=['subject_id','hadm_id','stay_id', 'charttime'], ascending=[True,True,True,True])
after_icu_data.reset_index(drop=True, inplace=True)

data = 0
for bidl in bidl_ls:
    sdata = after_icu_data[after_icu_data['hadm_id']==bidl]
    if len(sdata)!=0:
        sdata_index = min(sdata.index)
        inputdata = before_icu_data_latest[before_icu_data_latest['hadm_id']==bidl].reset_index(drop=True)

        sliced_inputdata = inputdata.loc[0, 'SpO2':'Lactate']
        after_icu_data.loc[sdata_index, 'SpO2':'Lactate'] = sliced_inputdata

after_icu_data['charttime'] = pd.to_datetime(after_icu_data['charttime'])
after_icu_data['outtime'] = pd.to_datetime(after_icu_data['outtime'])
after_icu_data['intime'] = pd.to_datetime(after_icu_data['intime'])

def resampling_func(x):
        column = ['subject_id','hadm_id','stay_id', 'time'] + feature_name[:14]

        feature = feature_name[:14]

        x.reset_index(drop=True, inplace=True)
        date_range = pd.date_range(x['intime'][0], x['outtime'][0], freq='10T', inclusive='right')
        new_df = pd.DataFrame(index=date_range, columns=column)
        
        new_df['subject_id'] = x['subject_id'][0]
        new_df['hadm_id'] = x['hadm_id'][0]
        new_df['stay_id'] = x['stay_id'][0]
        new_df['time'] = new_df.index
        
        for i in range(len(date_range)):
                if i == 0:
                    condition1 = date_range[i]>=x['charttime']
                    sa = x[condition1]
                    for p in feature:
                        new_df[p][i] = sa[p].mean()
                else :
                    condition2 = date_range[i-1]<x['charttime']
                    condition3 = date_range[i]>=x['charttime']
                    sa = x[(condition2)&(condition3)]

                    for p in feature:
                        new_df[p][i] = sa[p].mean()
        return new_df

result = after_icu_data.groupby('stay_id').apply(resampling_func)
result = result.reset_index(drop=True)
print('---------------------------------------------------------------------------------')
result.to_csv('./10min vital lab Dataset.csv', index=False)

result.columns

print(result.shape)
result.isnull().sum()
